+++ runtime/lib/native_calls.c
@@ -221,7 +221,7 @@ CL_Object* cl_call_native(CL_NativeData* data, CL_FUNC_PARAMS) {
     return ret_obj;
 }
 
-CL_Object* cl_native(const char* func, const char* library, enum CL_NativeType result_type, unsigned int count, ...) {
+CL_Object* cl_native(const char* func, const char* library, enum CL_NativeType result_type, size_t count, ...) {
     va_list args;
     va_start(args, count);
 
@@ -235,7 +235,7 @@ CL_Object* cl_native(const char* func, const char* library, enum CL_NativeType r
         data->args_types = NULL;
     }
 
-    for (unsigned int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         enum CL_NativeType type = va_arg(args, enum CL_NativeType);
         data->args_types[i] = type;
     }
