+++ runtime/lib/objects/list.h
@@ -14,7 +14,7 @@ bool cl_is_list_internal(CL_Object* obj);
 
 CL_Object* cl_list_at(CL_FUNC_PARAMS);
 
-unsigned int cl_list_length_internal(CL_Object* obj);
+size_t cl_list_length_internal(CL_Object* obj);
 
 CL_Object* cl_list_length(CL_FUNC_PARAMS);
